Move package utilities to utils.py
