+++ src/eval_framework/tasks/registry.py
@@ -5,11 +5,10 @@ from collections.abc import Generator, Iterator, Sequence
 from typing import Annotated, Any
 
 import pydantic
-from packaging.requirements import Requirement
-from packaging.version import Version
 from pydantic import AfterValidator
 
 from eval_framework.tasks.base import BaseTask
+from eval_framework.utils import is_extra_installed, validate_package_extras
 
 __all__ = [
     "register_task",
@@ -24,55 +23,10 @@ __all__ = [
 ]
 
 
-def _validate_package_extras(extras: str | Sequence[str], /, *, package: str = "eval_framework") -> Sequence[str]:
-    """Validate that the specified extras are valid for the given package."""
-    if isinstance(extras, str):
-        extras = [extras]
-
-    metadata = importlib.metadata.metadata(package)
-    package_extras = set(metadata.get_all("Provides-Extra") or [])
-    for extra in extras:
-        if extra not in package_extras:
-            raise ValueError(f"Invalid extra: {extra}. Options are {package_extras}")
-
-    return extras
-
-
 def validate_import_path(import_path: str) -> str:
     if importlib.util.find_spec(import_path) is None:
         raise ValueError(f"Invalid import path: {import_path}")
     return import_path
-
-
-def extra_requires(extra: str, /, *, package: str = "eval_framework") -> list[str]:
-    _validate_package_extras(extra, package=package)
-    dist = importlib.metadata.distribution(package)
-    requires = dist.requires or []
-    extra_str = f"extra == '{extra}'"
-    return [r.split(";")[0].strip() for r in requires if r.endswith(extra_str)]
-
-
-def _dependency_satisfied(dep: str, /) -> bool:
-    """Return True if the dependency string is satisfied.
-
-    Args:
-        A dependency string: for example "torch~=2.0".
-    """
-    try:
-        dist = importlib.metadata.distribution(Requirement(dep).name)
-        installed_version = Version(dist.version)
-        req = Requirement(dep)
-        return installed_version in req.specifier
-    except (importlib.metadata.PackageNotFoundError, Exception):
-        return False
-
-
-def is_extra_installed(extra: str, package: str = "eval_framework") -> bool:
-    """Return `True` if"""
-    for req in extra_requires(extra, package=package):
-        if not _dependency_satisfied(req):
-            return False
-    return True
 
 
 class TaskPlaceholder(pydantic.BaseModel, extra="forbid", frozen=True):
@@ -88,7 +42,7 @@ class TaskPlaceholder(pydantic.BaseModel, extra="forbid", frozen=True):
     extras: Annotated[
         tuple[str, ...],
         "Extra dependencies that are required for the task",
-        AfterValidator(_validate_package_extras),
+        AfterValidator(validate_package_extras),
     ] = ()
 
     def load(self) -> type[BaseTask]:
